services/api/requirements.txt
requests
fastapi
uvicorn
psycopg2-binary
python-dotenv
rapidfuzz